Moved alias provider and append functionality from Select*Appender to ExpressionSetAppender to unify handling of aliasing and appending across all assembly parts.  Fixed issues with override of equals, incorrect implementations on EntityExpression and FieldExpression.
