+++ src/HatTrick.DbEx.Sql/Expression/EntityExpression.cs
@@ -93,7 +93,9 @@ namespace HatTrick.DbEx.Sql.Expression
 
         public static bool operator ==(EntityExpression obj1, EntityExpression obj2)
         {
-            if (ReferenceEquals(null, obj1) && !ReferenceEquals(null, obj2)) return false;
+            if (ReferenceEquals(obj1, obj2)) return true;
+            if (ReferenceEquals(obj1, null)) return false;
+            if (ReferenceEquals(obj2, null)) return false;
             return obj1.Equals(obj2);
         }
