+++ src/HatTrick.DbEx.Sql/Expression/_FieldExpressions/FieldExpression.cs
@@ -140,7 +140,9 @@ namespace HatTrick.DbEx.Sql.Expression
 
         public static bool operator ==(FieldExpression obj1, FieldExpression obj2)
         {
-            if (ReferenceEquals(null, obj1) && !ReferenceEquals(null, obj2)) return false;
+            if (ReferenceEquals(obj1, obj2)) return true;
+            if (ReferenceEquals(obj1, null)) return false;
+            if (ReferenceEquals(obj2, null)) return false;
             return obj1.Equals(obj2);
         }
