+++ src/HatTrick.DbEx.Sql/HatTrick.DbEx.Sql.csproj
@@ -71,6 +71,7 @@
     <Compile Include="Assembler\ISqlParameterBuilderFactory.cs" />
     <Compile Include="Assembler\EnumValueTypeFormatter.cs" />
     <Compile Include="Assembler\ValueTypeFormatter.cs" />
+    <Compile Include="Assembler\_Appenders\ExpressionSetAppender.cs" />
     <Compile Include="Assembler\_Appenders\LiteralAppender.cs" />
     <Compile Include="Assembler\_Appenders\_Functions\CastFunctionAppender.cs" />
     <Compile Include="Assembler\_Appenders\_Functions\ConcatFunctionAppender.cs" />
@@ -100,7 +101,6 @@
     <Compile Include="Assembler\_Appenders\JoinAppender.cs" />
     <Compile Include="Assembler\_Appenders\SelectAppender.cs" />
     <Compile Include="Assembler\ISqlParameterBuilder.cs" />
-    <Compile Include="Assembler\SelectAllSqlStatementAssembler.cs" />
     <Compile Include="Assembler\SelectSqlStatementAssembler.cs" />
     <Compile Include="Assembler\SqlStatementBuilderFactory.cs" />
     <Compile Include="Assembler\SqlParameterBuilder.cs" />
